+++ src/components/ContextProvider/index.js
@@ -3,10 +3,7 @@ import React, {
   Component,
 } from 'react';
 
-
-import {
-  Context,
-} from "../../App";
+import Context from '@prisma-cms/context';
 
 import * as UI from "../ui"
 
@@ -59,8 +56,14 @@ class CooperationContextProvider extends Component {
 
   }
 
-
   prepareQuery() {
+
+    return {
+      ...this.prepareUserQuery(),
+    }
+  }
+
+  prepareUserQuery() {
 
 
     const {