+++ src/components/pages/Tasks/View/Gantt/DataTask/index.js
@@ -341,4 +341,6 @@ class DataTaskCustom extends PrismaCmsDataTask {
 // }
 
 // export default processors(withStyles(styles)(DataTaskCustom));
-export default withStyles(styles)(DataTaskCustom);
+export default withStyles(styles)(props => <DataTaskCustom 
+  {...props}
+/>);