+++ src/components/ui/Link/index.js
@@ -49,4 +49,6 @@ export class UiLink extends Component {
 }
 
 
-export default withStyles(styles)(UiLink);
+export default withStyles(styles)(props => <UiLink 
+  {...props}
+/>);