+++ src/dev/Renderer/MainMenu/User/index.js
@@ -1,7 +1,7 @@
 import React, { Component, Fragment } from 'react';
 
 import PropTypes from 'prop-types';
-import { Context } from '../../../../App';
+import Context from "@prisma-cms/context";
 
 
 // import { Timer } from "../../../components/ui";